The R Markdown blastula_email output format

The R Markdown blastula_email output format

blastula_email(
  content_width = "1000px",
  toc = FALSE,
  toc_depth = 3,
  toc_float = FALSE,
  number_sections = FALSE,
  section_divs = TRUE,
  fig_width = 5.35,
  fig_height = 5,
  fig_retina = 2,
  fig_caption = TRUE,
  dev = "png",
  smart = TRUE,
  self_contained = TRUE,
  template = "blastula",
  includes = NULL,
  keep_md = FALSE,
  md_extensions = NULL,
  connect_footer = FALSE,
  ...
)

Arguments

content_width

The width of the rendered HTML content. By default, this is set to 1000px. Using widths less than 600px is generally not advised but, if necessary, be sure to test such HTML emails with a wide range of email clients before sending to the intended recipients. The Outlook mail client (Windows, Desktop) does not respect content_width.

toc

If you would like an automatically-generated table of contents in the output email, choose TRUE. By default, this is FALSE where no table of contents will be generated.

toc_depth

The depth of headers to include in the table of contents ( should toc be set to TRUE).

toc_float

An option to float the table of contents to the left of the main document content. By default, this is FALSE.

number_sections

Sections can be sequentially numbered if this is set to TRUE. By default, this is FALSE.

section_divs

This wraps sections in <section> tags and attaches identifiers to the enclosing <section>s. This is set to TRUE.

fig_width, fig_height

The figure width and height in units of inches.

fig_retina

The scaling factor for retina displays. The default value is 2, which is the preferred choice for most retina displays. This can be set to NULL to prevent retina scaling. Note that this will always be NULL if keep_md is set to TRUE.

fig_caption

An option to render figures with captions. By default, this is set to TRUE.

dev

The R graphics device for figures. By default, this is the png device.

smart

An option to produce typographically correct output. This will convert straight quotes to curly quotes, --- to em dashes, -- to en dashes, and instances of ... to ellipses. By default, this is TRUE.

self_contained

Should a self-contained output file be generated. By default, this is TRUE. The standalone HTML file will have no external dependencies, it will use URIs to incorporate the contents of linked scripts, stylesheets, images, and videos.

template

The Pandoc template to use for rendering. This is the "blastula" template by default.

includes

A named list of additional content to include within the document. This is typically created using the rmarkdown::includes() function. By default, this is set to NULL.

keep_md

Should you need the keep the intermediate Markdown (.md) file, set this to TRUE. By default, the .md file is not kept.

md_extensions

Markdown extensions to be added or removed from the default definition or R Markdown.

connect_footer

Should a prepared footer message with links be included in the rendered email?

...

Specify other options in rmarkdown::html_document().