Default template for compose_email()

A template function that is suitable for using as the template argument of compose_email(). Template functions should generally not be called directly. When implementing your own template function, you must include parameters for html_body, html_header, html_footer, and title; you may also optionally add your own parameters, which callers to compose_email() can provide through the ... argument.

blastula_template(
  html_body,
  html_header,
  html_footer,
  title,
  content_width = "1000px",
  font_family = "Helvetica, sans-serif"
)

Arguments

html_body, html_header, html_footer
htmltools tag objects (e.g. htmltools::tags() or htmltools::HTML()), or NULL to omit.

title
Plain text title to be used for the <title> element; may be displayed in mobile phone notifications.

content_width
The width that should be used for the content area. By default, this is set to 1000px. Using widths less than 600px is generally not advised but, if necessary, be sure to test such HTML emails with a wide range of email clients before sending to the intended recipients.

font_family
The CSS value to use for font-family.

Value

A string containing a complete HTML document.