Prepare a email test message object

Create an email test message object, which is helpful for sending a test message with the send_email_out() function.

prepare_test_message(incl_ggplot = FALSE, incl_image = FALSE)

Arguments

incl_ggplot
an option to include a ggplot plot within the body of the test message. This requires that the ggplot2 package is installed. By default, this is FALSE.

incl_image
an option to include a test image within the body of the test message. By default, this is FALSE.

Examples

# NOT RUN {
# Create a credentials file to
# send via Gmail (this will be named
# `.bls_smtp_gmail_com`)
create_email_creds_file(
  user = "username@gmail.com",
  password = "*************",
  provider = "gmail",
  sender = "Sender Name")

# Send oneself a test message to
# test these new SMTP settings and also
# to ensure that the message appears
# correctly in the email client
send_email_out(
  message = prepare_test_message(),
  sender = "Sender Name",
  subject = "test 2",
  from = "username@gmail.com",
  to = "username@gmail.com",
  creds_file = ".bls_smtp_gmail_com")
# }