R Markdown render functions for the blastula_email output format

The render_email() and render_connect_email() functions both allow for rendering an an email message. We can supply an R Markdown document (.Rmd) with the output specified as output: blastula::blastula_email. While the render_email() and render_connect_email() functions have similar arguments, the render_connect_email() is preferred when publishing to the RStudio Connect service. It allows for the inclusion of a predefined footer that contains useful links for email recipients.

render_email(
  input,
  envir = parent.frame(),
  quiet = TRUE,
  output_options = list(),
  render_options = list()
)

render_connect_email(
  input,
  connect_footer = TRUE,
  envir = parent.frame(),
  quiet = TRUE,
  output_options = list(),
  render_options = list()
)

Arguments

input
The input file to be rendered. This should be an R Markdown document (.Rmd) with the output specified as output: blastula::blastula_email.

envir
The environment in which the code chunks are to be evaluated during knitting.

quiet
An option to suppress printing of the command line output from Pandoc during rendering. By default, this is set to TRUE.

output_options, render_options
Lists of options can be used to augment the rendering of the email message. The output_options list will be passed as the output_options argument of rmarkdown::render(). The render_options list is for providing additional arguments to rmarkdown::render(). By default, both lists are empty.

connect_footer
Should a prepared footer message with links be included in the rendered email? This argument is only available in the render_connect_email() function and is set to TRUE by default.