Send an email message through SMTP

Send an email message to one or more recipients via an SMTP server.

smtp_send(email, from, to, subject = NULL, cc = NULL, bcc = NULL,
  creds_file = NULL, sender = NULL, host = NULL, port = NULL,
  user = NULL, password = NULL, use_ssl = TRUE,
  authenticate = TRUE, binary_loc = NULL, echo = FALSE,
  echo_cmd = FALSE)

Arguments

email
The email message object, as created by the compose_email() function. The object's class is email_message

from
The email address of the sender. This does not have to be the same email that is associated with the account actually sending the message.

to
A vector of email addresses serving as primary recipients for the message. For secondary recipients, use the cc and bcc arguments.

subject
The subject of the message, which is usually a brief summary of the topic of the message.

cc
A vector of email addresses for sending the message as a carbon copy. This list of for those who are to receive a copy of a message addressed primarily to another. The list of recipients in the CC list is visible to all other recipients of the message.

bcc
A vector of email addresses for sending the message as blind carbon copies. Any email addresses provided here will receive the message and these email addresses will be concealed from other recipients (including others on the BCC list).

creds_file
An optional path to an email credentials file. This file must be created by the create_email_creds_file() function.

sender
The sender name.

host
The email host.

port
The port associated with the email account.

user
The username associated with the email account.

password
The password associated with the email account.

use_ssl
A logical value to indicate whether to use SSL.

authenticate
A logical value to indicate whether to use authentication.

echo
An option to print the standard output and error to the screen.

echo_cmd
A logical value indicating whether the system command should be printed to the console during the sending of email.

Examples

# NOT RUN {
# Prepare a test message and send
# the email out with `smtp_send()`
prepare_test_message() %>%
  smtp_send(
    from = "sender@mail.com",
    to = "recipient@mail.com",
    subject = "Mail Subject",
    creds_file = ".mail_creds"
  )
# }