Send an email message through SMTP

Send an email message to one or more recipients via an SMTP server. The email message required as input to smtp_send() has to be created by using the compose_email() function. The email_message object can be previewed by printing the object, where the HTML preview will show how the message should appear in recipients' email clients. We can avoid re-entering SMTP configuration and credentials information by retrieving this information either from disk (with the file generated by use of the create_smtp_creds_file() function), or, from the system's key-value store (with the key set by the create_smtp_creds_key() function).

smtp_send(email, to, from, subject = NULL, cc = NULL, bcc = NULL,
  credentials = NULL, binary_loc = NULL, echo = FALSE,
  dry_run = FALSE, creds_file = "deprecated")

Arguments

email
The email message object, as created by the compose_email() function. The object's class is email_message.

to
A vector of email addresses serving as primary recipients for the message. For secondary recipients, use the cc and bcc arguments.

from
The email address of the sender. Often this needs to be the same email address that is associated with the account actually sending the message.

subject
The subject of the message, which is usually a brief summary of the topic of the message.

cc
A vector of email addresses for sending the message as a carbon copy. This list of for those who are to receive a copy of a message addressed primarily to another. The list of recipients in the CC list is visible to all other recipients of the message.

bcc
A vector of email addresses for sending the message as blind carbon copies. Any email addresses provided here will receive the message and these email addresses will be concealed from other recipients (including others on the BCC list).

credentials
One of three credential helper functions must be used here: (1) creds(), (2) creds_key(), or (3) creds_file(). The first, creds(), allows for a manual specification of SMTP configuration and credentials within that helper function. This is the most secure method for supplying credentials as they aren't written to disk. The creds_key() function is used if credentials are stored in the system-wide key-value store, through use of the create_smtp_creds_key() function. The creds_file() helper function relies on a credentials file stored on disk. Such a file is created using the create_smtp_creds_file() function.

binary_loc
An option to supply the location of the mailsend-go binary file should it not be on the system path or in the working directory.

echo
If set to TRUE, the command for sending the message via mailsend-go will be printed to the console. By default, this is FALSE.

dry_run
Setting dry_run to TRUE will return information on the SMTP sending options. Furthermore, the function will stop short of actually sending the email message out. By default, however, this is set to FALSE.

creds_file
An option to specify a credentials file. As this argument is deprecated, please consider using credentials = creds_file(<file>) instead.

Details

To send messages, we need access to the mailsend-go binary, which is cross-platform and works on Windows, macOS (via Homebrew), and Linux (Debian and RPM packages). Instructions for installation can be found at https://github.com/muquit/mailsend-go#downloading-and-installing.

Examples

# NOT RUN {
# Before sending out an email through
# SMTP, we need an `email_message`
# object; for the purpose of a simple
# example, we can use the function
# `prepare_test_message()` to create
# a test version of an email (although
# we'd normally use `compose_email()`)
email <- prepare_test_message()

# The `email` message can be sent
# through the `smtp_send()` function
# so long as we supply the appropriate
# credentials; The following three
# examples provide scenarios for both
# the creation of credentials and their
# retrieval within the `credentials`
# argument of `smtp_send()`

# (1) Providing the credentials info
# directly via the `creds()` helper
# (the most secure means of supplying
# credentials information)

email %>%
  smtp_send(
    from = "sender@email.com",
    to = "recipient@email.com",
    credentials = creds(
      provider = "gmail",
      user = "sender@email.com")
  )

# (2) Using a credentials key (with
# the `create_smtp_creds_key()` and
# `creds_key()` functions)

create_smtp_creds_key(
 id = "gmail",
 user = "sender@email.com",
 provider = "gmail"
 )

email %>%
  smtp_send(
    from = "sender@email.com",
    to = "recipient@email.com",
    credentials = creds_key(
      "gmail"
      )
  )

# (3) Using a credentials file (with
# the `create_smtp_creds_file()` and
# `creds_file()` functions)

create_smtp_creds_file(
 file = "gmail_secret",
 user = "sender@email.com",
 provider = "gmail"
 )

email %>%
  smtp_send(
    from = "sender@email.com",
    to = "recipient@email.com",
    credentials = creds_file(
      "gmail_secret")
  )
# }